:bug: (neo) create project pkg contain readme field
