+++ .nvmrc
@@ -1,1 +1,1 @@
-v14.17.3
+v14.19.1
